requirements.txt
# Audio handling
sounddevice==0.4.6
soundfile==0.12.1
numpy>=1.24.0
scipy>=1.10.0
pyaudio==0.2.13

# Speech recognition
RealtimeSTT==0.3.99

# Text-to-speech
TTS==0.8.0  # Coqui TTS client library
fastapi==0.68.1  # For TTS server
uvicorn==0.15.0  # For TTS server

# LLM Integration
requests>=2.31.0  # For API calls to Ollama

# Configuration
pyyaml>=6.0.1

# Development
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0

# Testing
pytest>=7.3.1
pytest-cov>=4.1.0 